+++ inst/examples/sample_mallows_example.R
@@ -31,18 +31,3 @@ model_fit <- compute_mallows(samples, nmc = 10000)
 # The highest posterior density interval covers alpha0 = 10.
 compute_posterior_intervals(model_fit, burnin = 2000, parameter = "alpha")
 
-# The PerMallows package has a Gibbs sampler for sampling from the Mallows
-# distribution with Cayley, Kendall, Hamming, and Ulam distances. For these
-# distances, using the PerMallows package is typically faster.
-
-# Let us sample 100 rankings from the Mallows model with Cayley distance,
-# with the same consensus ranking and scale parameter as above.
-library(PerMallows)
-# Set the scale parameter of the PerMallows package corresponding to
-# alpha0 in BayesMallows
-theta0 <- alpha0 / n_items
-# Sample with PerMallows::rmm
-sample1 <- rmm(n = 100, sigma0 = rho0, theta = theta0, dist.name = "cayley")
-# Generate the same sample with sample_mallows
-sample2 <- sample_mallows(rho0 = rho0, alpha0 = alpha0, n_samples = 100,
-                          burnin = 1000, thinning = 1000, metric = "cayley")